+++ src/SteadyWaves.jl
@@ -12,6 +12,8 @@ module SteadyWaves
 include("params.jl")
 using .Params
 
+include("index.jl")
+
 include("output.jl")
 using .Output
 using .Output: wave_height, wavelength, wave_power, wave_period
