+++ src/index.jl
@@ -1,0 +1,50 @@
+module Index
+"""
+`u[elevation_indexes(N)]`: free surface elevations *kη*
+"""
+function elevation_indexes(N)
+return 1:N+1
+end
+
+"""
+`u[stream_indexes(N)]` : stream function coefficients *B*
+"""
+function stream_indexes(N)
+return N+2:2N+1
+end
+
+"""
+`u[2N+C_INDEX]`: wave celerity *c√(k/g)*
+"""
+const C_INDEX::Int = 2
+
+"""
+`u[2N+D_INDEX]`: mean water depth *kη̄*
+"""
+const D_INDEX::Int = 3
+
+"""
+`u[2N+Q_INDEX]`: volume flux due to waves *q√(k³/g)*
+"""
+const Q_INDEX::Int = 4
+
+"""
+`u[2N+R_INDEX]`: Bernoulli constant *rk/g*
+"""
+const R_INDEX::Int = 5
+
+"""
+`u[2N+U_INDEX]`: mean flow velocity *Ū√(k/g)*
+"""
+const U_INDEX::Int = 6
+
+"""
+`u[2N+H_INDEX]`: wave height *kH*
+"""
+const H_INDEX::Int = 7
+
+export C_INDEX, D_INDEX, H_INDEX, Q_INDEX, R_INDEX, U_INDEX
+
+export elevation_indexes,stream_indexes
+
+end
